correction

+++ src/BmiService.java
@@ -1,7 +1,7 @@
 public class BmiService {
 
-    public double calculate(double height, double weight) {
+    public int calculate(double height, double weight) {
         double index = weight / (height * height);
-        return index;
+        return (int) index;
     }
 }
