+++ src/Main.java
@@ -4,7 +4,7 @@ public class Main {
         BmiService service = new BmiService();
         double height = 1.87;
         double weight = 98;
-        int index = (int) service.calculate(height, weight);
+        int index = service.calculate(height, weight);
         System.out.println(index);
     }
 }
